requirements.txt
asgiref==3.8.0
Django==5.0.3
django-jalali==6.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
jdatetime==4.1.1
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
